Move VBA parser into WordMacro
